refactor: remove BottomSheet component and use updated useBottomSheet hook
