+++ src/components/BottomSheet/index.ts
@@ -1,2 +1,1 @@
 export * from "./useBottomSheet";
-export * from "./BottomSheet";
